requirements.txt
audioread==3.0.1
joblib==1.4.2
librosa==0.11.0
matplotlib
mido==1.3.3
numpy
pandas
# PyAudio==0.2.14
pyfluidsynth==1.3.4
scikit-learn
scipy
